rename: Utils 클래스 racinggame.Utils로 패키지 변경
